1 [VER] 1.31.2 2 [CHG] add check if dir is empty when web host remove reference file

+++ main/product/Codes/DevLib.Web.Hosting.WebHost40/WebServer.cs
@@ -9,6 +9,7 @@ namespace DevLib.Web.Hosting.WebHost40
     using System.Diagnostics.CodeAnalysis;
     using System.Globalization;
     using System.IO;
+    using System.Linq;
     using System.Net;
     using System.Net.Sockets;
     using System.Reflection;
@@ -706,14 +707,27 @@ namespace DevLib.Web.Hosting.WebHost40
 
             if (!this._isBinFolderExists)
             {
-                try
+                if (this.IsDirectoryEmpty(this._binFolder))
                 {
-                    Directory.Delete(this._binFolder, true);
-                }
-                catch
-                {
+                    try
+                    {
+                        Directory.Delete(this._binFolder, true);
+                    }
+                    catch
+                    {
+                    }
                 }
             }
+        }
+
+        /// <summary>
+        /// Determines whether the specified path is empty directory.
+        /// </summary>
+        /// <param name="sourcePath">The path to check.</param>
+        /// <returns>true if the specified path is empty directory; otherwise, false.</returns>
+        private bool IsDirectoryEmpty(string sourcePath)
+        {
+            return !Directory.EnumerateFileSystemEntries(sourcePath, "*", SearchOption.AllDirectories).Any();
         }
     }
 }
